requirements.txt
streamlit>=1.0
streamlit-image-comparison
Pillow
